Added a function that will generate a random shape anywhere on the screen

+++ app/src/main/java/apps/smoll/dragdropgame/Shape.kt
@@ -1,6 +1,6 @@
 package apps.smoll.dragdropgame
 
-data class Shape(val coordinates: Pair<Float, Float>, var typeResource: Int, val colorResource: Int)
+data class Shape(val coordinates: Pair<Int, Int>, var typeResource: Int, val colorResource: Int)
 
 enum class ShapeType {
     STAR, SQUARE, CIRCLE, HEXAGON
